backend/requirements.txt
flask
flask-sqlalchemy
flask_migrate
flask-cors
pymysql
werkzeug
boto3
PyJWT
Pillow
